Removed old attributes runtime module
